refactor(frontend): share one EnvironmentSelector across MCP and agent forms
The MCP catalog form and the agent dialog each had their own Environment
picker. They diverged: the MCP one showed each environment's description (in
the dropdown and below the select) and a "manage / only default available"
helper, while the agent one was a bare name-only dropdown.

Extract the MCP form's selector into a shared `EnvironmentSelector` that
self-fetches the environment list, permissions, and the org default, and use
it in both forms. The agent dialog now shows environment descriptions too;
it passes `hideWhenOnlyDefault` to keep hiding the field when no custom
environment is available, and supplies its card styling via `className`.
Both forms keep mapping the "Default" option to a null environmentId.
